adding support for tags on shortcuts
