+++ app/src/main/java/fr/neamar/kiss/pojo/Pojo.java
@@ -33,6 +33,59 @@ public abstract class Pojo {
     // Variable to store the formated (user selection in bold) tag
     public String displayTags = "";
 
+    // Tags normalized
+    public String tagsNormalized;
+    // Array that contains the non-normalized positions for every normalized
+    // character entry
+    private int[] tagsPositionMap = null;
+
+    public void setTags(String tags) {
+        // Set the actual user-friendly name
+        this.tags = tags;
+
+        if (this.tags != null) {
+            this.tags = this.tags.replaceAll("<", "&lt;");
+            // Normalize name for faster searching
+            Pair<String, int[]> normalized = StringNormalizer.normalizeWithMap(this.tags);
+            this.tagsNormalized = normalized.first;
+            this.tagsPositionMap = normalized.second;
+        }
+    }
+
+    public void setTagHighlight(int positionStart, int positionEnd) {
+        int posStart = this.mapTagsPosition(positionStart);
+        int posEnd = this.mapTagsPosition(positionEnd);
+
+        this.displayTags = this.tags.substring(0, posStart)
+                + '{' + this.tags.substring(posStart, posEnd) + '}' + this.tags.substring(posEnd, this.tags.length());
+    }
+//
+//    public void setTagHighlight(int positionStart, int positionEnd) {
+//        this.displayTags = this.tags.substring(0, positionStart)
+//                + '{' + this.tags.substring(positionStart, positionEnd) + '}'
+//                + this.tags.substring(positionEnd);
+//    }
+    /**
+     * Map a position in the normalized name to a position in the standard name string
+     *
+     * @param position Position in normalized name
+     * @return Position in non-normalized string
+     */
+    public int mapTagsPosition(int position) {
+        if (this.tagsPositionMap != null) {
+            if (position < this.tagsPositionMap.length) {
+                return this.tagsPositionMap[position];
+            }
+            return this.tags.length();
+        } else {
+            // No mapping defined
+            if (position < this.tags.length()) {
+                return position;
+            }
+            return this.tags.length();
+        }
+    }
+
     /**
      * Map a position in the normalized name to a position in the standard name string
      *
@@ -109,12 +162,6 @@ public abstract class Pojo {
             lastPositionEnd = positionEnd;
         }
         this.displayName += this.name.substring(lastPositionEnd);
-    }
-
-    public void setTagHighlight(int positionStart, int positionEnd) {
-        this.displayTags = this.tags.substring(0, positionStart)
-                + '{' + this.tags.substring(positionStart, positionEnd) + '}'
-                + this.tags.substring(positionEnd);
     }
     
     /**
