+++ app/src/main/res/layout/item_shortcut.xml
@@ -1,13 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<RelativeLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:gravity="center_vertical"
     android:orientation="horizontal"
     android:paddingBottom="@dimen/result_margin_bottom"
+    android:paddingEnd="@dimen/result_margin_right"
     android:paddingLeft="@dimen/result_margin_left"
     android:paddingRight="@dimen/result_margin_right"
+    android:paddingStart="@dimen/result_margin_left"
     android:paddingTop="@dimen/result_margin_top">
 
     <ImageView
@@ -21,21 +24,34 @@
         android:contentDescription="@null"
         tools:src="@drawable/ic_launcher" />
 
-    <TextView
-        android:id="@+id/item_app_name"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignBottom="@+id/item_shortcut_icon"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentRight="true"
-        android:layout_alignTop="@+id/item_shortcut_icon"
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_centerVertical="true"
+        android:layout_toEndOf="@+id/item_shortcut_icon"
         android:layout_toRightOf="@+id/item_shortcut_icon"
-        android:ellipsize="end"
-        android:gravity="center_vertical"
-        android:singleLine="true"
-        android:text="@string/stub_application"
-        android:textColor="?attr/resultColor"
-        android:textSize="@dimen/result_title_size" />
+        android:orientation="vertical">
+
+        <TextView
+            android:id="@+id/item_app_name"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:ellipsize="end"
+            android:gravity="center_vertical"
+            android:singleLine="true"
+            android:text="@string/stub_application"
+            android:textColor="?attr/resultColor"
+            android:textSize="@dimen/result_title_size" />
+
+        <TextView
+            android:id="@+id/item_app_tag"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="-6dp"
+            android:ellipsize="end"
+            android:text="@string/stub_app_tag"
+            android:textColor="?android:attr/textColorSecondary"/>
+    </LinearLayout>
 
     <ImageView
         android:id="@+id/item_app_icon"
