+++ app/src/main/res/menu/menu_item_shortcut.xml
@@ -3,9 +3,15 @@
     <item
         android:id="@+id/item_favorites_add"
         android:title="@string/menu_favorites_add" />
+
+    <item
+        android:id="@+id/item_tags_edit"
+        android:title="@string/menu_tags_edit" />
+
     <item
         android:id="@+id/item_favorites_remove"
         android:title="@string/menu_favorites_remove" />
+
     <item
         android:id="@+id/item_app_uninstall"
         android:title="@string/menu_shortcut_remove" />
